# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pod Quiz Documentation",
    "description": "API documentation for Pod Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "http://localhost:9200"
    },
    {
      "url": "http://172.20.10.12:9200"
    }
  ],
  "paths": {
...
    "/api/restaurant": {
      "post": {
        "tags": [
          "Manage-Restaurants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create a new restaurant",
        "operationId": "postNewRestaurant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Restaurant ABC"
                  },
                  "description": {
                    "type": "string",
                    "example": "A cozy place with great food."
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string",
                        "example": "Tehran"
                      },
                      "province": {
                        "type": "string",
                        "example": "Tehran"
                      },
                      "postalCode": {
                        "type": "string",
                        "example": "875764"
                      },
                      "mapLocation": {
                        "type": "string",
                        "example": "86r7er67e5e6"
                      }
                    }
                  }
                },
                "required": [
                  "name",
                  "address"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restaurant added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Restaurant added successfully"
                    },
                    "newRestaurant": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109cb"
                        },
                        "name": {
                          "type": "string",
                          "example": "Restaurant ABC"
                        },
                        "description": {
                          "type": "string",
                          "example": "A cozy place with great food."
                        },
                        "address": {
                          "type": "object",
                          "properties": {
                            "city": {
                              "type": "string",
                              "example": "Tehran"
                            },
                            "province": {
                              "type": "string",
                              "example": "Tehran"
                            },
                            "postalCode": {
                              "type": "string",
                              "example": "875764"
                            },
                            "mapLocation": {
                              "type": "string",
                              "example": "86r7er67e5e6"
                            }
                          }
                        },
                        "isActive": {
                          "type": "boolean",
                          "example": true
                        },
                        "menu": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "60d0fe4f5311236168a109cc"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/restaurant/{restaurantId}": {
      "put": {
        "tags": [
          "Manage-Restaurants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update a restaurant",
        "operationId": "updateRestaurant",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Restaurant Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description."
                  },
                  "address": {
                    "type": "string",
                    "example": "456 Elm St, City, Country"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restaurant updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Restaurant updated successfully"
                    },
                    "updatedRestaurant": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Manage-Restaurants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Delete a restaurant",
        "operationId": "deleteRestaurant",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurant deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Restaurant deleted successfully"
                    },
                    "deleteRestaurant": {
                      "type": "string",
                      "example": "Restaurant ABC"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-access-token",
        "in": "header"
      }
    }
  }
}